+++ src/graphics/animation.rs
@@ -93,32 +93,49 @@ impl Animation {
         self.frame_length = new_frame_length;
     }
 
-    /// Get the current frame that is active. This can be used in combination with [frames](#method.frames) to track the progress of this animation.
+    /// Gets the index of the frame that is currently being displayed.
+    ///
+    /// This index is zero-based, and can be used in combination with the [`frames`](#method.frames)
+    /// method in order to track the progress of the animation.
     pub fn current_frame_index(&self) -> usize {
         self.current_frame
     }
 
-    /// Set the current frame that is active.  This can be used for fine-tuned animation manipulation.
+    /// Sets which frame of the animation should be displayed.
     ///
-    /// The given value should be a valid index in the [frames](#method.frames) list, otherwise this animation will panic.
-    pub fn set_current_frame_index(&mut self, new_frame_index: usize) {
-        // Without this check, the code would panic in `Drawable::draw` because `self.frames[self.current_frame]` is invalid,
-        // but the developer would have no clue where it was set.
-        debug_assert!(self.frames.get(new_frame_index).is_some());
+    /// Usually you will want to control the animation by calling [`advance`](#method.advance)
+    /// or [`advance_by`](#method.advance), but this method can be useful for more fine-grained
+    /// control.
+    ///
+    /// The index is zero-based, and must be within the bounds of the animation's
+    /// [`frames`](#method.frames), otherwise this method will panic.
+    pub fn set_current_frame_index(&mut self, index: usize) {
+        // Without this check, the code would panic in `Drawable::draw` because `self.frames[self.current_frame]`
+        // is invalid, but the developer would have no clue where it was set.
+        assert!(index < self.frames.len());
 
-        self.current_frame = new_frame_index;
+        self.current_frame = index;
     }
 
-    /// Get the duration that this frame has been visible. This can be used in combination with [frame_length](#method.frame_length) to track the progress of this animation.
+    /// Gets the duration that the current frame has been visible.
+    ///
+    /// This can be used in combination with the [`frame_length`](#method.frame_length) method
+    /// in order to track the progress of the animation.
     pub fn current_frame_time(&self) -> Duration {
         self.timer
     }
 
-    /// Set the duration that the current frame has been visible. This can be used for fine-tuned animation manipulation.
+    /// Sets the duration that the current frame has been visible.
     ///
-    /// The animation will not update until the next call to [advance](#method.advance) or [advance_by](#method.advance_by). If a value is given that is larger than [frame_length](#method.frame_length), this animation may skip frames.
-    pub fn set_current_frame_time(&mut self, new_frame_time: Duration) {
-        self.timer = new_frame_time;
+    /// Usually you will want to control the animation by calling [`advance`](#method.advance)
+    /// or [`advance_by`](#method.advance), but this method can be useful for more fine-grained
+    /// control.
+    ///
+    /// The animation will not advance past the end of the current frame until the next call
+    /// to [`advance`](#method.advance) or [`advance_by`](#method.advance_by). If a value is
+    /// given that is larger than [`frame_length`](#method.frame_length), this animation may skip frames.
+    pub fn set_current_frame_time(&mut self, duration: Duration) {
+        self.timer = duration;
     }
 }